+++ DalaMock.Sample/WindowService.cs
@@ -21,21 +21,25 @@ public class WindowService : IHostedService
     private readonly IDalamudPluginInterface pluginInterface;
     private readonly IUiBuilder uiBuilder;
     private readonly SampleWindow sampleWindow;
+    private readonly DtrBarSampleWindow dtrBarSampleWindow;
     private readonly IWindowSystem windowSystem;
 
-    public WindowService(WindowSystemFactory windowSystemFactory, IDalamudPluginInterface pluginInterface, IUiBuilder uiBuilder, SampleWindow sampleWindow)
+    public WindowService(WindowSystemFactory windowSystemFactory, IDalamudPluginInterface pluginInterface, IUiBuilder uiBuilder, SampleWindow sampleWindow, DtrBarSampleWindow dtrBarSampleWindow)
     {
         this.windowSystemFactory = windowSystemFactory;
         this.pluginInterface = pluginInterface;
         this.uiBuilder = uiBuilder;
         this.sampleWindow = sampleWindow;
+        this.dtrBarSampleWindow = dtrBarSampleWindow;
         this.windowSystem = this.windowSystemFactory.Create("DalaMock.Sample");
         this.windowSystem.AddWindow(sampleWindow);
+        this.windowSystem.AddWindow(dtrBarSampleWindow);
     }
 
     public Task StartAsync(CancellationToken cancellationToken)
     {
         this.sampleWindow.IsOpen = true;
+        this.dtrBarSampleWindow.IsOpen = true;
         this.uiBuilder.Draw += this.Draw;
         return Task.CompletedTask;
     }