+++ DalaMock/Mocks/MockPluginLog.cs
@@ -1,3 +1,5 @@
+using Microsoft.Extensions.Logging;
+
 namespace DalaMock.Core.Mocks;
 
 using System;
@@ -7,9 +9,9 @@ using Serilog.Events;
 
 public class MockPluginLog : IPluginLog, IMockService
 {
-    public readonly ILogger Logger;
+    public ILogger<MockPluginLog> Logger { get; }
 
-    public MockPluginLog(ILogger logger)
+    public MockPluginLog(ILogger<MockPluginLog> logger)
     {
         this.Logger = logger;
     }
@@ -23,7 +25,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Fatal(messageTemplate, values);
+        this.Logger.LogCritical(messageTemplate, values);
     }
 
     public void Fatal(Exception? exception, string messageTemplate, params object[] values)
@@ -33,7 +35,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Fatal(exception, messageTemplate, values);
+        this.Logger.LogCritical(exception, messageTemplate, values);
     }
 
     public void Error(string messageTemplate, params object[] values)
@@ -43,7 +45,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Error(messageTemplate, values);
+        this.Logger.LogError(messageTemplate, values);
     }
 
     public void Error(Exception? exception, string messageTemplate, params object[] values)
@@ -53,7 +55,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Error(exception, messageTemplate, values);
+        this.Logger.LogError(exception, messageTemplate, values);
     }
 
     public void Warning(string messageTemplate, params object[] values)
@@ -63,7 +65,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Warning(messageTemplate, values);
+        this.Logger.LogWarning(messageTemplate, values);
     }
 
     public void Warning(Exception? exception, string messageTemplate, params object[] values)
@@ -73,7 +75,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Warning(exception, messageTemplate, values);
+        this.Logger.LogWarning(exception, messageTemplate, values);
     }
 
     public void Information(string messageTemplate, params object[] values)
@@ -83,7 +85,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Information(messageTemplate, values);
+        this.Logger.LogInformation(messageTemplate, values);
     }
 
     public void Information(Exception? exception, string messageTemplate, params object[] values)
@@ -93,7 +95,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Information(exception, messageTemplate, values);
+        this.Logger.LogInformation(exception, messageTemplate, values);
     }
 
     public void Info(string messageTemplate, params object[] values)
@@ -103,7 +105,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Information(messageTemplate, values);
+        this.Logger.LogInformation(messageTemplate, values);
     }
 
     public void Info(Exception? exception, string messageTemplate, params object[] values)
@@ -113,7 +115,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Information(exception, messageTemplate, values);
+        this.Logger.LogInformation(exception, messageTemplate, values);
     }
 
     public void Debug(string messageTemplate, params object[] values)
@@ -123,7 +125,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Debug(messageTemplate, values);
+        this.Logger.LogDebug(messageTemplate, values);
     }
 
     public void Debug(Exception? exception, string messageTemplate, params object[] values)
@@ -133,7 +135,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Debug(exception, messageTemplate, values);
+        this.Logger.LogDebug(exception, messageTemplate, values);
     }
 
     public void Verbose(string messageTemplate, params object[] values)
@@ -143,7 +145,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Verbose(messageTemplate, values);
+        this.Logger.LogTrace(messageTemplate, values);
     }
 
     public void Verbose(Exception? exception, string messageTemplate, params object[] values)
@@ -153,7 +155,7 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Verbose(exception, messageTemplate, values);
+        this.Logger.LogTrace(exception, messageTemplate, values);
     }
 
     public void Write(LogEventLevel level, Exception? exception, string messageTemplate, params object[] values)
@@ -163,7 +165,32 @@ public class MockPluginLog : IPluginLog, IMockService
             return;
         }
 
-        this.Logger.Write(level, exception, messageTemplate, values);
+        LogLevel logLevel = LogLevel.Trace;
+
+        switch (level)
+        {
+            case LogEventLevel.Verbose:
+                break;
+            case LogEventLevel.Debug:
+                logLevel = LogLevel.Debug;
+                break;
+            case LogEventLevel.Information:
+                logLevel = LogLevel.Information;
+                break;
+            case LogEventLevel.Warning:
+                logLevel = LogLevel.Warning;
+                break;
+            case LogEventLevel.Error:
+                logLevel = LogLevel.Error;
+                break;
+            case LogEventLevel.Fatal:
+                logLevel = LogLevel.Critical;
+                break;
+            default:
+                throw new ArgumentOutOfRangeException(nameof(level), level, null);
+        }
+
+        this.Logger.Log(logLevel, exception, messageTemplate, values);
     }
 
     public LogEventLevel MinimumLogLevel { get; set; } = LogEventLevel.Verbose;