+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [2.0.25] - 2024-08-04
+
+### Added
+
+- Added IDtrBar mock service
+
+### Fixed
+
+- Use Microsoft.Extensions.Logging internally in combination with serilog
+- Updated Serilog to latest stable release
+
 ## [2.0.24] - 2024-08-04
 
 ### Fixed